requirements.txt
asgiref==3.8.1
Django==5.0.4
sqlparse==0.5.0
tzdata==2024.1
djangorestframework==3.14.0
psycopg==3.1.18
python-decouple==3.8
setuptools
pytest==7.4.2
pytest-django==4.5.2
drf-yasg==1.21.5
djangorestframework-simplejwt==5.2.2
django-log-request-id==1.7.0
django-cors-headers==3.14.0
django-filter==23.2
black==23.3.0
flake8==6.1.0
isort==5.12.0
